AP_Baro: improved barometer averaging
this changes the barometer calculations to floating point. On a MS5611
this is actually about twice as fast as the previous 64 bit
calculations, but gains us more accuracy as we are able to take
advantage of sub-bit precision when we average over 8 samples.

+++ libraries/AP_Baro/AP_Baro_BMP085.cpp
@@ -128,11 +128,11 @@ uint8_t AP_Baro_BMP085::read()
 	return(result);
 }
 
-int32_t AP_Baro_BMP085::get_pressure() {
+float AP_Baro_BMP085::get_pressure() {
     return Press;
 }
 
-int16_t AP_Baro_BMP085::get_temperature() {
+float AP_Baro_BMP085::get_temperature() {
     return Temp;
 }
